(MSVC) silence warnings

+++ gfx/drivers/d3d12.c
@@ -109,7 +109,7 @@ static void d3d12_gfx_sync(d3d12_video_t* d3d12)
 #ifdef HAVE_OVERLAY
 static void d3d12_free_overlays(d3d12_video_t* d3d12)
 {
-   int i;
+   size_t i;
    for (i = 0; i < (unsigned)d3d12->overlays.count; i++)
       d3d12_release_texture(&d3d12->overlays.textures[i]);
 
@@ -188,7 +188,7 @@ static void d3d12_overlay_set_alpha(void* data, unsigned index, float mod)
 
 static bool d3d12_overlay_load(void* data, const void* image_data, unsigned num_images)
 {
-   int i;
+   size_t i;
    D3D12_RANGE range;
    d3d12_sprite_t*             sprites = NULL;
    d3d12_video_t*              d3d12   = (d3d12_video_t*)data;
@@ -288,7 +288,7 @@ static void d3d12_get_overlay_interface(void* data, const video_overlay_interfac
 
 static void d3d12_render_overlay(d3d12_video_t *d3d12)
 {
-   int i;
+   size_t i;
 
    if (d3d12->flags & D3D12_ST_FLAG_OVERLAYS_FULLSCREEN)
    {
@@ -510,7 +510,7 @@ static void d3d12_update_viewport(d3d12_video_t *d3d12, bool force_full)
 
 static void d3d12_free_shader_preset(d3d12_video_t* d3d12)
 {
-   int i;
+   size_t i;
    if (!d3d12->shader_preset)
       return;
 
@@ -611,7 +611,7 @@ static void d3d12_init_pipeline(
 static bool d3d12_gfx_set_shader(void* data, enum rarch_shader_type type, const char* path)
 {
 #if defined(HAVE_SLANG) && defined(HAVE_SPIRV_CROSS)
-   int i;
+   size_t i;
    d3d12_texture_t* source = NULL;
    d3d12_video_t*   d3d12  = (d3d12_video_t*)data;
 
@@ -1962,7 +1962,7 @@ error:
 
 static void d3d12_init_history(d3d12_video_t* d3d12, unsigned width, unsigned height)
 {
-   int i;
+   size_t i;
    /* TODO/FIXME: should we init history to max_width/max_height instead ?
     * to prevent out of memory errors happening several frames later
     * and to reduce memory fragmentation */
@@ -1982,7 +1982,7 @@ static void d3d12_init_history(d3d12_video_t* d3d12, unsigned width, unsigned he
 
 static void d3d12_init_render_targets(d3d12_video_t* d3d12, unsigned width, unsigned height)
 {
-   int i;
+   size_t i;
    for (i = 0; i < d3d12->shader_preset->passes; i++)
    {
       struct video_shader_pass* pass = &d3d12->shader_preset->pass[i];
